make c++ ts work up to arrays
