+++ rmw_iceoryx_cpp/CMakeLists.txt
@@ -42,7 +42,11 @@ find_package(rosidl_typesupport_introspection_cpp REQUIRED)
 
 add_library(rmw_iceoryx_serialization SHARED
   src/internal/iceoryx_deserialize.cpp
+  src/internal/iceoryx_deserialize_typesupport_c.cpp
+  src/internal/iceoryx_deserialize_typesupport_cpp.cpp
   src/internal/iceoryx_serialize.cpp
+  src/internal/iceoryx_serialize_typesupport_c.cpp
+  src/internal/iceoryx_serialize_typesupport_cpp.cpp
 )
 target_include_directories(rmw_iceoryx_serialization
   PUBLIC include
@@ -142,9 +146,15 @@ if(BUILD_TESTING)
   ament_lint_auto_find_test_dependencies()
 
   find_package(ament_cmake_gtest REQUIRED)
-  ament_add_gtest(${PROJECT_NAME}_tests test/iceoryx_name_conversion_test.cpp)
+  find_package(test_msgs REQUIRED)
+  ament_add_gtest(test_name_conversion test/iceoryx_name_conversion_test.cpp)
+  target_link_libraries(test_name_conversion ${PROJECT_NAME})
 
-  target_link_libraries(${PROJECT_NAME}_tests ${PROJECT_NAME})
+  ament_add_gtest(test_message_serialization test/iceoryx_serialization_test.cpp)
+  target_link_libraries(test_message_serialization ${PROJECT_NAME})
+  ament_target_dependencies(test_message_serialization
+    test_msgs
+  )
 endif()
 
 ament_export_include_directories(include)
